backend/requirements.txt
Flask>=3.1.2
Flask-Cors>=6.0.1
Flask-JWT-Extended>=4.7.1
bcrypt>=5.0.0
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1
requests>=2.32.4
tweepy>=4.14.0
google-genai>=1.0.0